+++ torchqdynamics/mesolve/rouchon.py
@@ -6,7 +6,7 @@ import torch
 import torch.nn as nn
 from torch import Tensor
 
-from ..ode.ode_adjoint_solver import ODEAdjointSolver
+from ..ode.adjoint_solver import ODEAdjointSolver
 from ..utils.solver_utils import inv_sqrtm, kraus_map
 from ..utils.utils import trace
